requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
Flask==1.1.2
numpy==1.19.5
torch==1.8.1+cpu 
torchvision==0.9.1+cpu
nltk==3.6.2
nlpaug==1.1.4
gunicorn==19.8.1
requests==2.18.4
extaugment==1.3.4